refactor: migrate heightmap-editor to bundled TypeScript controller
Port public/modules/ui/heightmap-editor.js (1713 lines) to
src/controllers/heightmap-editor.ts following the migration guide (strict
d3 v7: select/drag/pointer/range/hsl/lab/leastIndex/interpolateRound named
imports, event-arg brush handlers, d3.scan→leastIndex). Covers all tools:
mode selection, brush painting, fill/line brushes, template editor DSL,
image converter, preview, and the erase/keep/risk finalize-and-regenerate
paths. The static #customizationMenu markup stays in index.html; panel
listeners wire once behind init flags. Register as Controllers.HeightmapEditor
and replace editHeightmap() at the heightmap-selection/tools/hotkeys call
sites; drop the editHeightmap global. Add color/edits/undraw/rankCells/
generatePrecipitation/changeViewMode/resetZoom/RgbQuant globals and Culture
x/y transient fields.
