+++ src/index.html
@@ -8570,7 +8570,6 @@
     <script defer src="modules/ui/editors.js?v=1.133.0"></script>
     <script defer src="modules/ui/tools.js?v=1.133.0"></script>
     <script defer src="modules/ui/world-configurator.js?v=1.131.0"></script>
-    <script defer src="modules/ui/heightmap-editor.js?v=1.131.0"></script>
     <script defer src="modules/ui/temperature-graph.js?v=1.106.6"></script>
     <script defer src="modules/ui/coastline-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/ai-generator.js?v=1.120.5"></script>
